+++ Sources/KeyboardKit/KeyboardInputViewController.swift
@@ -19,6 +19,14 @@ import UIKit
  keyboard extension with a lot of additional features that a
  regular keyboard extension lacks.
  
+ Regarding the view controller lifecycle, there are some new
+ functions that you can override. These should be overridden
+ to assure that the keyboard is kept in sync, e.g.:
+ 
+ * `viewWillSetupKeyboard` is called when a controller needs
+ to setup or recreate the keyboard view. Override this, then
+ call `setup(with:)` or any other setup logic.
+ 
  This class provides you with many utils that you can use to
  build a more powerful keyboard extension, for instance:
  
@@ -65,6 +73,18 @@ open class KeyboardInputViewController: UIInputViewController {
     open override func viewWillLayoutSubviews() {
         keyboardContext.sync(with: self)
         super.viewWillLayoutSubviews()
+        viewWillSyncWithScreenSize()
+    }
+    
+    open func viewWillSetupKeyboard() {
+        // Override and implement your keyboard setup logic.
+    }
+    
+    open func viewWillSyncWithScreenSize() {
+        let size = UIScreen.main.bounds.size
+        if size == lastScreenSize { return }
+        lastScreenSize = size
+        viewWillSetupKeyboard()
     }
     
     open override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
@@ -172,6 +192,13 @@ open class KeyboardInputViewController: UIInputViewController {
     public var textInputProxy: TextInputProxy? {
         didSet { keyboardContext.sync(with: self) }
     }
+    
+    /**
+     This internal property is used to listen to screen size
+     changes. Since iOS 15, keyboard views must be recreated
+     whenever the screen size changes.
+     */
+    var lastScreenSize: CGSize = .zero
     
     
     // MARK: - Observables